requirements.txt
Django==1.11.4
django-ipware==1.1.6
pytz==2017.2
wheel==0.24.0
python-dateutil==2.6.1
